backend/requirement.txt
astroid==2.5
attrs==21.4.0
blinker==1.4
cachelib==0.6.0
certifi==2021.10.8
chardet==4.0.0
click==8.0.4
Deprecated==1.2.13
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==2.10
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyJWT==1.6.4
pylint==2.6.0
pylint-venv==2.1.1
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==6.2.2
redis==4.1.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.12.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2

